add create account unit test

+++ db/sqlc/accounts/query.sql.go
@@ -19,7 +19,7 @@ INSERT INTO accounts (
 
 type CreateAccountParams struct {
 	Owner    string
-	Balance  int64
+	Balance  float64
 	Currency string
 }
 
@@ -114,7 +114,7 @@ RETURNING id, owner, balance, currency, created_at
 
 type UpdateAccountParams struct {
 	ID      int64
-	Balance int64
+	Balance float64
 }
 
 func (q *Queries) UpdateAccount(ctx context.Context, arg UpdateAccountParams) error {
